factored out card from carousel component using React 'Containment' strategy. Deleted old Carousel component files.
