+++ client/src/components/RelatedItemsAndOutfit/UserOutfits/UserOutfit.jsx
@@ -1,11 +1,21 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import OutfitCarousel from './OutfitCarousel';
+import Carousel from './Carousel';
+import OutfitListCard from './OutfitListCard';
 
 const UserOutfit = ({ prodsInfo, prodsStyles, prodsMeta }) => (
-  <div>
+  <div className="outfit-list-container">
     Related Products Here:
-    <OutfitCarousel prodsInfo={prodsInfo} prodsStyles={prodsStyles} prodsMeta={prodsMeta} />
+    <Carousel prodsInfo={prodsInfo} prodsStyles={prodsStyles} prodsMeta={prodsMeta} >
+      {prodsInfo.map((prodInfo, index) => (
+        <OutfitListCard
+          key={prodInfo.id}
+          prodInfo={prodInfo}
+          prodStyles={prodsStyles[index]}
+          prodMeta={prodsMeta[index]}
+        />
+      ))}
+    </Carousel>
   </div>
 );
